Message receiving enabled
This application can now receive messages. One thread is in charge of sending messages, and the other thread is in charge of receiving messages. Though one concern I have is that I use the exit() function to terminate this program. When the user enters in the termination string, only the first thread—the thread that sends messages—terminates; the other thread doesn't terminate. I'm not entirely sure how to make it so that when the first thread terminates, the other thread will instantly terminate as well. Well, not without using exit() at least.